# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zyvor Fleet API
  version: 0.3.0
  description: Offline-first edge fleet control-plane API with dynamic targeting, health-gated staged rollouts and rollback
    safety.
servers:
- url: /
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Signed in. Sets HttpOnly session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error'
  /api/v1/auth/logout:
    post:
      tags:
      - Auth
      security:
      - cookieAuth: []
      parameters:
      - $ref: '#/components/parameters/ZyvorRequest'
      responses:
        '200':
          description: Signed out
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zyvor_fleet_session
    bearerAuth:
      type: http
      scheme: bearer
    apiTokenAuth:
      type: http
      scheme: bearer
      description: Scoped Fleet API token with zf_api_ prefix
  parameters:
...
    ZyvorRequest:
      name: X-Zyvor-Request
      in: header
      required: true
      schema:
        type: string
        const: '1'
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              status:
                type: integer
  schemas:
    Role:
      type: string
      enum:
      - admin
      - operator
      - viewer
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
...